Style+Widgets: .h => .hpp
