+++ src/libHeaven/Widgets/ModeSwitchWidget.cpp
@@ -19,8 +19,8 @@
 
 #include "libHeaven/App/Mode.hpp"
 
-#include "libHeaven/Widgets/ModeSwitchWidget.h"
-#include "libHeaven/Widgets/ModeSwitchCombo.h"
+#include "libHeaven/Widgets/ModeSwitchWidget.hpp"
+#include "libHeaven/Widgets/ModeSwitchCombo.hpp"
 
 namespace Heaven
 {
